requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==24.3.0
bigquery-magics==0.5.0
blinker==1.9.0
branca==0.8.1
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dash==2.18.2
dash_ag_grid==31.3.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash_daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
db-dtypes==1.3.1
debugpy==1.8.11
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
fastavro==1.10.0
fiona==1.10.1
Flask==3.0.3
folium==0.19.2
fonttools==4.55.3
frozenlist==1.5.0
geopandas==0.14.4
google-api-core==2.24.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.11.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.22.0
google-cloud-storage==2.19.0
google-crc32c==1.1.2
google-genai==0.3.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
gunicorn==23.0.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
libcst==1.5.1
mapclassify==2.8.1
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
multidict==6.1.0
munkres==1.1.4
mypy_extensions==1.0.0
nest_asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==24.3.1
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.28.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydata-google-auth==1.9.0
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.0
pyproj==3.7.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyu2f==0.1.5
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.3
rsa==4.9
Rtree==1.3.0
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
shapely==2.0.6
six==1.17.0
stack_data==0.6.3
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
typing_inspect==0.9.0
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
xyzservices==2024.9.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
